refactor(ViewHelper) move HTML tag to ComponentMount
