+++ lib/react/rails.rb
@@ -2,5 +2,6 @@ require 'react/rails/asset_variant'
 require 'react/rails/engine'
 require 'react/rails/railtie'
 require 'react/rails/version'
+require 'react/rails/component_mount'
 require 'react/rails/view_helper'
 require 'react/rails/controller_renderer'
